Move cors middleware to a new file
